Fixing a few tests

+++ test/src/me/anno/tests/utils/SphericalHierarchyTest.kt
@@ -5,6 +5,7 @@ import me.anno.utils.assertions.assertEquals
 import org.joml.Matrix4x3
 import org.joml.Matrix4x3d
 import org.joml.Vector3d
+import org.joml.Vector3f
 import org.junit.jupiter.api.Test
 
 class SphericalHierarchyTest {
@@ -15,12 +16,12 @@ class SphericalHierarchyTest {
         val c = Vector3d(+2.0, +0.0, 1.0)
         val tri = SphereTriangle(null, 0, 0, a, b, c)
         // strategic tests
-        assertEquals(Vector3d(0.0, 1.0, 0.0), tri.baseAB)
-        assertEquals(Vector3d(0.0, 0.0, 1.0), tri.baseUp)
-        assertEquals(Vector3d(1.0, 0.0, 0.0), tri.baseAC)
-        assertEquals(tri.baseAB, tri.localToGlobal.transformDirection(Vector3d(1.0, 0.0, 0.0)))
-        assertEquals(tri.baseUp, tri.localToGlobal.transformDirection(Vector3d(0.0, 1.0, 0.0)))
-        assertEquals(tri.baseAC, tri.localToGlobal.transformDirection(Vector3d(0.0, 0.0, 1.0)))
+        assertEquals(Vector3f(0f, 1f, 0f), tri.baseAB)
+        assertEquals(Vector3f(0f, 0f, 1f), tri.baseUp)
+        assertEquals(Vector3f(1f, 0f, 0f), tri.baseAC)
+        assertEquals(tri.baseAB, tri.localToGlobal.transformDirection(Vector3f(1f, 0f, 0f)))
+        assertEquals(tri.baseUp, tri.localToGlobal.transformDirection(Vector3f(0f, 1f, 0f)))
+        assertEquals(tri.baseAC, tri.localToGlobal.transformDirection(Vector3f(0f, 0f, 1f)))
         assertEquals(tri.globalToLocal, tri.localToGlobal.invert(Matrix4x3()))
         assertEquals(tri.localA, tri.globalToLocal.transformPosition(a, Vector3d()))
         assertEquals(tri.localB, tri.globalToLocal.transformPosition(b, Vector3d()))
